+++ ui.R
@@ -53,7 +53,7 @@ ui <- fluidPage(
     tags$link(rel = "stylesheet", type = "text/css", href = "style.css"),
     includeHTML("google-analytics.html"),
     tags$script(src = "html2canvas.min.js"),
-    tags$script(src = "saveAs.js"),
+    tags$script(src = "saveAs.js")
   ),
   useShinyjs(),